Calculate long/short liquidation payouts
When calculating the payout combinations based on our payout curve
implementation and the parameters provided we originally only defined
the so-called settlement payouts. This is the set of payout
combinations which is to be used for:

- Constructing the CETs which enable the non-collaborative settlement
  of the CFD based on oracle attestations which occur at the time of
  expiry of the CFD.
- Settling the CFD collaboratively before its expiry.

We want to introduce another mechanism to close the CFD _unilaterally_
before expiry. We refer to this feature as dynamic liquidation and it
is based on liquidation payouts.

In this patch we simply calculate the payouts to be used when
liquidation occurs. We choose the very tails of the payout curve for
simplicity.