Abstract over oracle actor in rollover protocol
The introduction of the `oracle::AnnouncementsChannel` is not strictly
necessary yet. But it will be once we extract rollover protocol into a
separate crate, so that we don't run into a situation where we are
trying to implement a foreign trait `rollover::GetAnnouncements` on a
foreign type `xtra::MessageChannel<_, _>`.

We may be able to find a simpler alternative if we don't use a trait
as part of the API of the rollover protocol.