Replace rollover confirmation step with configurable flag
This simplifies both the maker's side of the rollover protocol and the
maker itself.

We were also using the confirmation step to provide certain offer
parameters back to the protocol execution. Having removed the step, we
nowfetch them from the `maker::cfd::Actor` via the new
`xtra_libp2p_rollover::GetRates` trait.

We remove two tests which no longer make sense because we now do not
pause the execution of the rollover protocol to get confirmation via
the HTTP API.